add fix for cygwin warning

+++ test_hm_file.c
@@ -92,7 +92,7 @@ void processfile(HM *table, const char *filename, int op){
 char *trim(char *s){
 	size_t i;
 	for(i = strlen(s); i > 0; i--)
-		if (isspace(s[i]) || s[i] == '\0')
+		if (isspace((int) s[i]) || s[i] == '\0')
 			s[i] = '\0';
 		else
 			break;
